Added new modules: admin and user
Added a superuser module of Admin which can read and delete any entry
Added a user module to read user info and change their password

+++ routers/todos.py
@@ -3,8 +3,10 @@ from starlette import status
 from pydantic import Field, BaseModel
 from models import Todos, Users
 from database import SessionLocal
+from sqlalchemy import select
 from sqlalchemy.orm import Session
 from typing import Annotated, Optional
+from .auth import get_current_user
 
 
 router = APIRouter(prefix="/todos", tags=["todos"])
@@ -19,6 +21,7 @@ def get_db():
         
         
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class newdata(BaseModel):
     title: Optional[str] = Field(None, min_length=3)
@@ -38,30 +41,48 @@ class newdata(BaseModel):
     
 #REST APIs for table: todos in the sqlmdb database
 
+
+
 @router.get("/all", status_code=status.HTTP_200_OK)
-def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency):
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="User authentication failed")
+    
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 @router.get("/{data_id}", status_code=status.HTTP_200_OK)
-def read_todos(db: db_dependency, data_id: int = Path(gt=0)):
-    data_model =  db.query(Todos).filter(Todos.id == data_id).first()
+async def read_todos(user: user_dependency, db: db_dependency, data_id: int = Path(gt=0)):
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="User authentication failed")
+    
+    data_model =  db.query(Todos).filter(Todos.owner_id == user.get("id"), Todos.id == data_id).first()
     if data_model:
         return data_model
     else:
         raise HTTPException(status_code=404, detail="Data not Found")
     
 @router.post("/new_todo",status_code=status.HTTP_201_CREATED)
-def create_todos(db: db_dependency, newdata1: newdata):
-    data_model = Todos(**newdata1.model_dump())
+async def create_todos(user: user_dependency, db: db_dependency, newdata1: newdata):
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="User authentication failed")
+    
+    data_model = Todos(**newdata1.model_dump(),owner_id = user.get("id"))
     
     db.add(data_model)
     db.commit()
     
 @router.put("/update_todo/{data_id}",status_code=status.HTTP_204_NO_CONTENT)
-def update_todos(db: db_dependency, data_id: int, newdata1: newdata):
-    data_model = db.query(Todos).filter(Todos.id == data_id).first()
+async def update_todos(user: user_dependency, db: db_dependency, data_id: int, newdata1: newdata):
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="User authentication failed")
+    
+    data_model = db.query(Todos).filter(Todos.owner_id == user.get("id"), Todos.id == data_id).first()
     if not data_model:
-        raise HTTPException(status_code=404, detail="Data not Found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Data not Found")
     
     if newdata1.title:
         data_model.title = newdata1.title
@@ -77,10 +98,14 @@ def update_todos(db: db_dependency, data_id: int, newdata1: newdata):
     
 
 @router.delete("/delete_todo/{data_id}",status_code=status.HTTP_204_NO_CONTENT)
-def delete_todos(db: db_dependency, data_id: int = Path(gt=0)):
-    data_model = db.query(Todos).filter(Todos.id == data_id).first()
+async def delete_todos(user: user_dependency, db: db_dependency, data_id: int = Path(gt=0)):
+    
+    if not user:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="User authentication failed")
+    
+    data_model = db.query(Todos).filter(Todos.owner_id == user.get("id"), Todos.id == data_id).first()
     if not data_model:
         raise HTTPException(status_code=404, detail="Data not Found")
     
-    db.query(Todos).filter(Todos.id == data_id).delete()
+    db.query(Todos).filter(Todos.owner_id == user.get("id"), Todos.id == data_id).delete()
     db.commit()
